chore: Move the Deserializer into the the serde module
