+++ meilidb-data/src/serde/mod.rs
@@ -1,0 +1,3 @@
+mod deserializer;
+
+pub use self::deserializer::Deserializer;
